move attribute logic
