+++ mocha.opts
@@ -1,2 +1,1 @@
---bail
 test/test.js
