+++ src/generators/nodes/Binding.ts
@@ -1,0 +1,7 @@
+import Node from './shared/Node';
+
+export default class Binding extends Node {
+	name: string;
+	value: Node[]
+	expression: Node
+}
