+++ src/generators/nodes/EventHandler.ts
@@ -1,0 +1,7 @@
+import Node from './shared/Node';
+
+export default class EventHandler extends Node {
+	name: string;
+	value: Node[]
+	expression: Node
+}
