+++ src/generators/nodes/Ref.ts
@@ -1,0 +1,7 @@
+import Node from './shared/Node';
+
+export default class Ref extends Node {
+	name: string;
+	value: Node[]
+	expression: Node
+}
