+++ src/generators/nodes/Transition.ts
@@ -1,0 +1,7 @@
+import Node from './shared/Node';
+
+export default class Transition extends Node {
+	name: string;
+	value: Node[]
+	expression: Node
+}
